DeleteSchedule

+++ FitnessUA/src/ua/nure/ostpc/malibu/shedule/server/ScheduleManagerServiceImpl.java
@@ -2220,15 +2220,22 @@ public class ScheduleManagerServiceImpl extends RemoteServiceServlet implements
 	public void removeSchedule(long id) 
 			throws IllegalArgumentException, OperationCallException {
 		Schedule s = nonclosedScheduleCacheService.getSchedule(id);
-		if (s.getPeriod().getStatus() != Status.DRAFT)
+		if (!(s.getPeriod().getStatus() == Status.DRAFT 
+				|| s.getPeriod().getStatus() == Status.FUTURE))
 			throw new OperationCallException(
 					"Данный график не имеет статус 'Черновик'");
-			try {
-				scheduleDAO.removeSchedule(id);
-			} catch (DAOException e) {
-				throw new IllegalArgumentException(
-						"Данный график не найден или сервер не доступен.");
-			}
+		
+		if (s.getPeriod().getPeriodId() != nonclosedScheduleCacheService.getLastSchedule().getPeriod().getPeriodId())
+			throw new OperationCallException(
+					"Вы не можете удалить не последний график");
+			
+		try {
+			scheduleDAO.removeSchedule(id);
+			nonclosedScheduleCacheService.removeSchedule(id);
+		} catch (DAOException e) {
+			throw new IllegalArgumentException(
+					"Данный график не найден или сервер не доступен.");
+		}
 	}
 
 	@Override
